requirements.txt
Django==3.2.13
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==22.1
django-mptt==0.14.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5



